Refactor of code structure to follow DDD onion architecture
