+++ carp_mobile_sensing/lib/infrastructure/data_managers/console_data_manager.dart
@@ -5,7 +5,7 @@
  * found in the LICENSE file.
  */
 
-part of '../data_managers.dart';
+part of '../infrastructure.dart';
 
 /// A very simple data manager that just "uploads" the data to the
 /// console (i.e., prints it). Used mainly for testing and debugging purposes.
