+++ carp_mobile_sensing/lib/infrastructure/data_managers/file_data_manager.dart
@@ -5,7 +5,7 @@
  * found in the LICENSE file.
  */
 
-part of '../data_managers.dart';
+part of '../infrastructure.dart';
 
 class FileDataManagerFactory implements DataManagerFactory {
   @override
