+++ carp_mobile_sensing/lib/infrastructure/file_study_manager.dart
@@ -5,7 +5,7 @@
  * found in the LICENSE file.
  */
 
-part of '../data_managers.dart';
+part of 'infrastructure.dart';
 
 /// Retrieve and store [StudyProtocol] json definitions on the device's local
 /// file system.
