+++ carp_mobile_sensing/lib/infrastructure/sampling_packages/device/device_data.dart
