+++ carp_mobile_sensing/lib/infrastructure/sampling_packages/sensors/light_probe.dart
