+++ carp_mobile_sensing/lib/runtime/device_manager.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: unnecessary_getters_setters
 
 /*
- * Copyright 2021-2022 Copenhagen Center for Health Technology (CACHET) at the
+ * Copyright 2021 Copenhagen Center for Health Technology (CACHET) at the
  * Technical University of Denmark (DTU).
  * Use of this source code is governed by a MIT-style license that can be
  * found in the LICENSE file.
