+++ carp_mobile_sensing/lib/services/notification_controller.dart
@@ -5,7 +5,7 @@
  * found in the LICENSE file.
  */
 
-part of '../runtime.dart';
+part of 'services.dart';
 
 /// A controller of user notifications allow for creating, scheduling, and
 /// canceling user notifications.
